src/Math/Calculator*.cs: Move generic version of Calculator class
... to its own file. This should increase the maintainability.

Resolves: No entry
